pyproject.toml
[project]
name = "telecom-assistant-hackathon"
version = "0.1.0"
description = "Telecom Assistant Multi-Agent System"
authors = [
    {name = "RakshithL"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "streamlit>=1.51.0,<2.0.0",
    "langgraph>=0.2.0,<1.0.0",
    # upgraded to LangChain 1.x to match langchain-community
    "langchain>=1.0.0,<2.0.0",
    "langchain-community>=0.4.1,<0.5.0",
    "crewai>=0.80.0,<1.0.0",
    "ag2>=0.10.0,<1.0.0",
    "llama-index>=0.12.0,<1.0.0",
    "pandas>=2.0.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "crewai-tools>=0.12.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "faiss-cpu>=1.7.0,<2.0.0",
    "llama-index-vector-stores-faiss>=0.1.0,<1.0.0"
]
